Put phase-specific prover code into its own modules
