+++ tlsn/tlsn-prover/src/tls/verify.rs
@@ -1,0 +1,7 @@
+//! This module handles the verification phase of the prover.
+//!
+//! The prover deals with a TLS verifier that is a notary and a verifier.
+
+use super::{state, Prover};
+
+impl Prover<state::Notarize> {}
